Improvement of grammar rules for better handling of C++; works for Atlas now
The notion of |function_head| was modified to include the initial (result) type
specifier, and is also used for |namespace| blocks. The rules we reorganised and
extended in several other aspects as well. The current Atlas code, with some
minor modifications, can be formatted acceptably. Such modifications involve
notably surrounding any function pointer identifier in its (possibly |typedef|)
definition, in addition to the obligatory parentheses, by @[ ... @]; this is a
necessary measure (mentioned in rules.w) to avoid wrong cweavex reduction.