More C++ adaptations: |function_head| reductions and rightward extensions
Since |function_head| could actually be a constructor application without
arguments, it reduces to |expression| when the context suggests this (for
instance a following comma; this used to reduce as for a declarator shared
between multiple defined identifiers, but that is unlikely for functions).
Also for class methods, not only can |const| follow, and be absorbed into, the
function head, the same goes for |&| and |&&| (categories |unorbinop|,|binop|).